requirements.txt
# Claude Skills Agentic Executor — pinned dependencies
# Generated from the installed environment.
# Install with:  pip install -r requirements.txt

# --- Core ---
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langchain-core==1.2.9
langchain-openai==1.1.7
langsmith==0.6.9
openai==2.17.0
pydantic==2.12.5
PyYAML==6.0.1

# --- SQLite checkpoint support ---
aiosqlite==0.22.1
sqlite-vec==0.1.6

# --- AI Provider SDKs ---
anthropic==0.52.0
google-genai==1.14.0

# --- Dev / Testing ---
pytest==9.0.2
pytest-asyncio==1.3.0
